fix compiler warnings and msvc opt options
mingw-gcc:
```
[132/150] compiling c muon.exe.p/src/platform/windows/run_cmd.c.o
../src/platform/windows/run_cmd.c: In function 'copy_pipes':
../src/platform/windows/run_cmd.c:137:87: warning: dereferencing type-punned pointer will break strict-aliasing rules [-Wstrict-aliasing]
  137 |                 if (!GetQueuedCompletionStatus(ctx->ioport, &_bytes_read, (PULONG_PTR)&pipe, &_overlapped, 100)) {
      |
```

msvc:
```
cl : Command line warning D9002 : ignoring unknown option '/0d'

../src/toolchains.c(1335): warning C4090: 'function': different 'const' qualifiers

../src/lang/analyze.c(1329): warning C4334: '<<': result of 32-bit shift implicitly converted to 64 bits (was 64-bit shift intended?)

../src/lang/eval.c(564): warning C4245: 'function': conversion from 'int' to 'uint32_t', signed/unsigned mismatch

../src/toolchains.c(1335): warning C4090: 'function': different 'const' qualifiers

D:\Downloads\muon\src\platform\windows\run_cmd.c(158) : warning C4754: Conversion rules for arithmetic operations in the comparison at D:\Downloads\muon\src\platform\windows\run_cmd.c(157) mean that one branch cannot be executed. Cast '(... * 2)' to 'ULONG64' (or similar type of 8 bytes).
```

+++ src/lang/eval.c
@@ -561,7 +561,7 @@ determine_build_file(struct workspace *wk, const char *cwd, enum build_language 
 	*out_lang = names[i].lang;
 
 	if (!quiet && *out_lang == build_language_cmake) {
-		vm_warning_at(wk, -1, "using experimental cmake compat mode, this will probably break");
+		vm_warning_at(wk, (uint32_t)-1, "using experimental cmake compat mode, this will probably break");
 	}
 
 	return get_cstr(wk, tstr_into_str(wk, &name));
